backend/requirements.txt
django==4.2.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.6
drf-yasg==1.21.7
python-decouple==3.8
ipython
